refactoring of reading notification, removing unused variables
