+++ src/routes/(app)/NotificationBell.svelte
@@ -2,7 +2,6 @@
   import Notification from "$lib/components/Notification.svelte";
   import type { Notification as NotificationType } from "@prisma/client";
   import { flip } from "svelte/animate";
-  import { invalidate } from "$app/navigation";
   import type { NotificationSchema } from "$lib/zod/schemas";
   import type { SuperValidated } from "sveltekit-superforms";
   import { superForm } from "sveltekit-superforms/client";
@@ -14,29 +13,9 @@
   // Get the number of unread notifications, which is then used to indicate the user
   $: unreadCount = notifications.filter((data) => data.readAt == null).length;
 
-  const readNotifications = async () => {
-    if (unreadCount <= 0) return;
-    // optimistic update
-    notifications = notifications.map((notification) => {
-      if (notification.readAt == null) {
-        notification.readAt = new Date();
-      }
-      return notification;
-    });
-    // send mutation request
-    try {
-      await fetch("/read-notifications", {
-        method: "PUT",
-      });
-      await invalidate("/notifications");
-    } catch (error) {
-      console.error("Could not read notifications", error); // do not show user, user did not take an explicit action so an error message would be confusing
-    }
-  };
-
   // this is a somewhat ugly way to keep focus on the bell button after a notification is removed
   // if this is not here, whenever an action is taken (such as removing a notification), the form will cause the elements to lose focus and the menu will close
-  let bellButton: HTMLDivElement;
+  let bellButton: HTMLButtonElement;
   const onUpdated = () => {
     bellButton.focus();
   };
@@ -57,26 +36,26 @@
 
 <div class="dropdown">
   <!-- When user clicks on the bell icon, it will update all notifications as read -->
-  <!-- svelte-ignore a11y-click-events-have-key-events -->
-  <div
-    bind:this={bellButton}
-    tabindex="0"
-    class="btn btn-ghost *:text-xl"
-    role="button"
-    data-dropdown-toggle="dropdown"
-    on:click={readNotifications}
-  >
-    {#if unreadCount <= 0}
-      <span class="i-mdi-bell-outline" />
-    {/if}
-    {#if unreadCount > 0}
-      <span class="i-mdi-bell-ring-outline animate-bounce" />
-      <span
-        class="absolute right-0 top-0 flex h-6 w-6 items-center justify-center rounded-box bg-red-600 text-center !text-sm"
-        >{unreadCount}</span
-      >
-    {/if}
-  </div>
+  <form method="POST" action="/notifications?/readNotification" use:enhance>
+    <button
+      bind:this={bellButton}
+      tabindex="0"
+      class="btn btn-ghost *:text-xl"
+      data-dropdown-toggle="dropdown"
+      type="submit"
+    >
+      {#if unreadCount <= 0}
+        <span class="i-mdi-bell-outline" />
+      {/if}
+      {#if unreadCount > 0}
+        <span class="i-mdi-bell-ring-outline animate-bounce" />
+        <span
+          class="absolute right-0 top-0 flex h-6 w-6 items-center justify-center rounded-box bg-red-600 text-center !text-sm"
+          >{unreadCount}</span
+        >
+      {/if}
+    </button>
+  </form>
   <!-- svelte-ignore a11y-no-noninteractive-tabindex -->
   <ul
     tabindex="0"
@@ -91,7 +70,7 @@
       <!-- Deletes all notifications -->
       <form
         method="POST"
-        action="/?/deleteNotification"
+        action="/notifications?/deleteNotification"
         use:enhance
         data-sveltekit-keepfocus
       >
