+++ src/routes/+layout.server.ts
@@ -4,8 +4,8 @@ import { loadFlash } from "sveltekit-flash-message/server";
  * Load the form flash message.
  * Propagates the user and member to the page data.
  */
-export const load = loadFlash(async ({ locals, depends }) => {
-  const { user, member, prisma } = locals;
+export const load = loadFlash(async ({ locals }) => {
+  const { user, member } = locals;
   return {
     user,
     member,
